Further simplify the from external merge
